Fix issue #8. Running translate-shell.vim on Windows.
The problem was connected because on Windows we are not able to check
if file is executable in custom shell (e.g we have set variable shell).

So, if user tries to run it on windows and trans file isn't executable
then we try to run it with help flag and check the output. If output is
empty then binary file doesn't exists.